All example tests use the new format
