+++ tests/unit/arch/riscv/test_example_003.mts
