+++ tests/unit/arch/riscv/test_example_004.mts
